restructuring files into a verticalslice architecture
